Add files via upload
Numerically controlled digital fabrication (NC) encompasses precise toolpaths generated according to the predefined geometry, material characteristics, type of machining processes, and processing tools. This includes a definition of tolerances, necessary for a successful and smooth assembly process when dealing with component-driven assembly after the components were fabricated. However, the prototyping and assembly practice often encounters a variety of issues in real fabrication and assembly scenarios: humidity of the environment, specific material properties, such as the ability of the stock material to change its dimensions according to the unexpected environmental conditions, unexpected different dimensions of the stock material delivered from a factory or a producer, variety of fabrication producers with different machine types, or unique conditions of a production machine. Therefore, there is a need to adapt the geometry of components to different humidity scenarios and combinations of stock material properties or specific machine or production-related processes, which influence the dimensions, and the tolerances embedded into the process of toolpath generation. To address these, we introduce an adaptive parametric model. This has been created as an open-source and open-access algorithm within the Grasshopper environment to easily modify the dimensions of tolerances according to specific and unexpected conditions.